Update 8-1.txt according to Errata
In the Errata for Exercise 8-1
(https://man7.org/tlpi/errata/index.html#p_166),
the author pointed out that the code of the original exercise wouldn't
yield the results described in the text.
Sorry that I assumed the text to be right and didn't found that out by
eye nor write a test to verify.